set ConnectionFsm and SystemFsm active

+++ Connection.c
@@ -21,16 +21,15 @@
 #include "base/Reactor.h"
 #include "base/AD.h"
 
+#include "SystemFsm.h"
 #include "Connection.h"
 #include "ConnectionContainer.h"
-
+#include "ConnectionFsm.h"
 #include "ServoDriver.h"
 
 #include "MessageSinkCdc.h"
 
 #define MAX_TIMEOUT_TIME 3
-
-typedef enum { CONN_IDLE, CONN_IDENTIFIED, CONN_UP, CONN_DEGRADED_1, CONN_DEGRADED_2, CONN_DEGRADED_3 } connectionState_t;
 
 typedef struct ConnectionContainer ConnectionContainer_t;
 
@@ -39,11 +38,11 @@ typedef struct Connection
   struct sockaddr_in srcAddress;
   int socketFd;
   char *pUserName;
-  connectionState_t state;
   UINT32 lastSequence;
   UINT32 timeout;
   void *hConnectionTimeoutPoll;
   ConnectionContainerObject_t connectionContainer;
+  ConnectionFsmObject_t connectionFsm;
   UINT8 hTrc; // trace handle
 } Connection_t;
 
@@ -52,48 +51,11 @@ typedef struct Connection
 #define CONNECTION_NOF_ACTION 2
 
 
-
-/**
- * @short Rocket Colibri protocol states, see state diagram
- */
-typedef enum
-{
-  CONNECTION_IDLE, // not active
-  CONNECTION_UP, //
-  CONNECTION_DEGRADED_1, //
-  CONNECTION_IDENTIFIED, //
-  CONNECTION_DEGRADED_2, //
-  CONNECTION_DEGRADED_3, //
-  CONNECTION_NOF_STATE
-} Connection_State_t;
-
-/**
- * @short List of all events
- */
-typedef enum
-{
-  CONNECTION_EVENT_START, // start state machine
-  CONNECTION_EVENT_OK, //
-  CONNECTION_EVENT_FAIL, //
-  CONNECTION_EVENT_STOP, //
-  CONNECTION_NOF_EVENT
-} Connection_Event_t;
-
-/** state machine action function signaure */
-typedef void (*Connection_ActionFn_t)(Connection_t *pArpingInst);
-
-/** connection event action */
-typedef struct
-{
-  Connection_ActionFn_t action[CONNECTION_NOF_ACTION + 1]; // +1 for NULL termination
-  Connection_State_t nextState;
-} Connection_EventAction_t;
-
 void HandleJsonMessage(ConnectionObject_t connectionObject, const char *pJsonString)
 {
   DBG_ASSERT(connectionObject);
   Connection_t *this = (Connection_t*) connectionObject;
-  this->timeout = 0; // reset timeout counter
+
   struct json_tokener *tok = json_tokener_new();
   struct json_object * jobj = json_tokener_parse_ex(tok, pJsonString, strlen(pJsonString));
   if (!jobj)
@@ -139,18 +101,27 @@ void HandleJsonMessage(ConnectionObject_t connectionObject, const char *pJsonStr
         	this->lastSequence = sequence;
           	TRC_INFO(this->hTrc, "cdc: user:%s sequence:%d", this->pUserName, this->lastSequence);
 
-          	ServoDriver_t *pServoDriver = ServoDriverGetInstance();
-            MessageSinkCdc_t *pMsgCdc = NewMessageSinkCdc(jobj);
-            pServoDriver->SetServos(GetNofChannel(pMsgCdc), GetChannelVector(pMsgCdc));
-            DeleteMessageSinkCdc(pMsgCdc);
+            if(SystemFsmIs_SYS_IDLE(ConnectionContainerGetSystemFsm(this->connectionContainer)))
+            {
+            	ConnectionFsmEventRecvCdcCmd(this->connectionFsm);
+            }
+            else if (SystemFsmIs_SYS_CONTROLLING(ConnectionContainerGetSystemFsm(this->connectionContainer)))
+			{
+              	ServoDriver_t *pServoDriver = ServoDriverGetInstance();
+                MessageSinkCdc_t *pMsgCdc = NewMessageSinkCdc(jobj);
+                pServoDriver->SetServos(GetNofChannel(pMsgCdc), GetChannelVector(pMsgCdc));
+                DeleteMessageSinkCdc(pMsgCdc);
+                this->timeout = 0; // reset timeout counter
+			}
+            else
+            {
+            	DBG_MAKE_ENTRY(FALSE);
+            }
           }
           else if (0 == strcmp(json_object_get_string(command), "hello"))
           {
-            // TODO
-          }
-          else if (0 == strcmp(json_object_get_string(command), "handover"))
-          {
-            // TODO
+        	  ConnectionFsmEventRecvHelloCmd(this->connectionFsm);
+        	  this->timeout = 0; // reset timeout counter
           }
           else
           {
@@ -188,13 +159,36 @@ void HandleJsonMessage(ConnectionObject_t connectionObject, const char *pJsonStr
     TIMERFD_Read(timerfd);
     if (this->timeout > MAX_TIMEOUT_TIME)
     {
-    	 TRC_ERROR(this->hTrc, "timeout expired %s", this->pUserName);
-    	ConnectionContainerRemoveConnection(this->connectionContainer, obj);
-    	DeleteConnection(obj);
+    	ConnectionFsmEventTimeout(this->connectionFsm);
+    	TRC_ERROR(this->hTrc, "timeout expired %s", this->pUserName);
     }
     else
+    {
+    	TRC_ERROR(this->hTrc, "this:0x%x timeout:%d",this, this->timeout);
     	this->timeout++;
+    }
   }
+
+static void A1_ActionFunctionSendToActive(ConnectionFsmObject_t obj)
+{
+	Connection_t *pConnection = (Connection_t *)ConnectionFsmGetConnection(obj);
+	ConnectionContainerSetActiveConnection(pConnection->connectionContainer, pConnection);
+	SystemFsmEventTransitionToActive(ConnectionContainerGetSystemFsm(pConnection->connectionContainer));
+}
+
+static void A2_ActionFunctionSendToPassive(ConnectionFsmObject_t obj)
+{
+	Connection_t *pConnection = (Connection_t *)ConnectionFsmGetConnection(obj);
+	ConnectionContainerSetActiveConnection(pConnection->connectionContainer, NULL);
+	SystemFsmEventTransitionToPassive(ConnectionContainerGetSystemFsm(pConnection->connectionContainer));
+}
+
+static void A3_ActionDeleteConnection(ConnectionFsmObject_t obj)
+{
+	Connection_t *pConnection = (Connection_t *)ConnectionFsmGetConnection(obj);
+	ConnectionContainerRemoveConnection(pConnection->connectionContainer, pConnection);
+	DeleteConnection(pConnection);
+}
 
   ConnectionObject_t NewConnection(const ConnectionContainerObject_t containerConnectionObject,
       const struct sockaddr_in *pSrcAddr, const int socketFd,  UINT8 hTrcSocket)
@@ -206,10 +200,15 @@ void HandleJsonMessage(ConnectionObject_t connectionObject, const char *pJsonStr
     this->connectionContainer = containerConnectionObject;
     this->srcAddress = *pSrcAddr;
     this->socketFd = socketFd;
-    this->state = CONN_IDLE;
     this->hTrc = hTrcSocket; // inherits the trace handle from the socket
     this->hConnectionTimeoutPoll = Reactor_AddReadFd(TIMERFD_Create(1000 * 1000),
         ConnectionTimeoutHandler, this, "ConnectionTimeoutRx");
+
+    this->connectionFsm = NewConnectionFsm(
+		this,
+    	A1_ActionFunctionSendToActive,
+    	A2_ActionFunctionSendToPassive,
+    	A3_ActionDeleteConnection);
 
     return (ConnectionObject_t)this;
   }
@@ -218,6 +217,7 @@ void HandleJsonMessage(ConnectionObject_t connectionObject, const char *pJsonStr
  {
 	 Connection_t *this = (Connection_t *) pConn;
 	 Reactor_RemoveFd(this->hConnectionTimeoutPoll);
+	 DeleteConnectionFsm(this->connectionFsm);
 	 free(pConn);
  }
   // getter
